requirements-test.txt
-r requirements-tox.txt
djangorestframework==3.13.1
django==4.0.4
pytest==7.1.2
